Replace new bspwm widget with python version
Because `bspc query` does not allow monitor names such as `DP-2.1.0`
which my docking station returns, we need a different way of accessing
this. Supposedly the selector `%<invalid monitor name>` is suppored, but
it simply doesn't work.

Instead parse the bspwm dump and assemble the widget in python (for
easier handling of the bspwm dump output)